get rid of some warnings in Windows build
For example:
db\client.cpp(632) : warning C4804: '>=' : unsafe use of type 'bool' in operation
db\pipeline\value.cpp(277): warning C4244: 'return' : conversion from 'const __int64' to 'double', possible loss of data

+++ db/pipeline/value.cpp
@@ -274,7 +274,7 @@ namespace mongo {
         if (type == NumberInt)
             return simple.intValue;
         if (type == NumberLong)
-            return simple.longValue;
+            return static_cast< double >( simple.longValue );
 
         assert(type == NumberDouble);
         return simple.doubleValue;
